Split up stepper indirection
